Changed type

+++ app/src/main/java/com/listener/musicplayerapp/domain/model/Song.kt
@@ -5,5 +5,5 @@ data class Song(
     val id: Int,
     val songName: String,
     val author: String,
-    val duration: Int
+    val duration: Long
 )
